openmetadata/satellite/requirements.txt
openmetadata-ingestion==1.1.5
flask